src/requirements.txt
django
mysqlclient
django-debug-toolbar
djangorestframework
drf-nested-routers
django-filter
djoser
djangorestframework_simplejwt
uwsgi
pillow
django-cors-headers
django-templated-mail